+++ src/App.svelte
@@ -13,6 +13,7 @@
   import TextToolsBar from './components/TextToolsBar.svelte';
   import CustomSpeedEstimator from './components/CustomSpeedEstimator.svelte';
   import AudioProofreader from './components/AudioProofreader.svelte';
+  import VocabularyRichness from './components/VocabularyRichness.svelte';
   import TextDiffViewer from './components/TextDiffViewer.svelte';
   import ReadabilityScore from './components/ReadabilityScore.svelte';
   import KeywordDensity from './components/KeywordDensity.svelte';
@@ -37,6 +38,8 @@
   <CustomSpeedEstimator />
 
   <AudioProofreader />
+
+  <VocabularyRichness />
 
   <TextDiffViewer />